pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.11",
    "groq>=0.19.0",
    "jinja2>=3.1.6",
    "langchain>=0.3.20",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.39",
    "tavily-python>=0.5.1",
    "uvicorn>=0.34.0",
    "python-dateutil>=2.9.0.post0",
    "httpx>=0.28.1",
    "pypdf>=5.3.1",
    "langchain-groq>=0.2.5",
]
